Fix a bunch of Python 3.10 issues with pure-Python classes and demos
In Python 3.10, a change[1] was implemented where extension functions
that take integer arguments will no longer silently accept non-integer
arguments (e.g., floats) that can only be converted to integers with a
loss of precision.  This PR fixes most of these issues in the pure-Python
classes and demos by explicitly converting the parameters to int before
passing them to wxWidgets.  There is loss of precision, but this was
happening before (automatically) anyway as most wxWidgets DeviceContext
functions operate using integers.

Additionally, the PR fixes a few sizing issues, mostly with SpinCtrls being
too small on GTK3.

This is an example of the relevant exception:
Traceback (most recent call last):
  File "/usr/lib64/python3.10/site-packages/wx/lib/agw/pygauge.py", line 355, in OnPaint
    r.width = w
TypeError: 'float' object cannot be interpreted as an integer

Fixes #2038.

[1] https://bugs.python.org/issue37999

(cherry picked from commit 173d0796810bb65de9bdfdc6941d24a04628f6c2)

+++ wx/lib/agw/balloontip.py
@@ -615,7 +615,7 @@ class BalloonTip(object):
         if delay < 1:
             raise Exception("\nERROR: Delay Time For BalloonTip Creation Should Be Greater Than 1 ms")
 
-        self._startdelaytime = float(delay)
+        self._startdelaytime = int(delay)
 
 
     def GetStartDelay(self):
@@ -640,7 +640,7 @@ class BalloonTip(object):
         if delay < 1:
             raise Exception("\nERROR: Delay Time For BalloonTip Destruction Should Be Greater Than 1 ms")
 
-        self._enddelaytime = float(delay)
+        self._enddelaytime = int(delay)
 
 
     def GetEndDelay(self):
