+++ wx/lib/plot/plotcanvas.py
@@ -173,24 +173,24 @@ class PlotCanvas(wx.Panel):
 
         # Default Pens
         self._gridPen = wx.Pen(wx.Colour(180, 180, 180, 255),
-                               self._pointSize[0],
+                               int(self._pointSize[0]),
                                wx.PENSTYLE_DOT)
 
         self._centerLinePen = wx.Pen(wx.RED,
-                                     self._pointSize[0],
+                                     int(self._pointSize[0]),
                                      wx.PENSTYLE_SHORT_DASH)
 
         self._axesPen = wx.Pen(wx.BLACK,
-                               self._pointSize[0],
+                               int(self._pointSize[0]),
                                wx.PENSTYLE_SOLID)
 
         self._tickPen = wx.Pen(wx.BLACK,
-                               self._pointSize[0],
+                               int(self._pointSize[0]),
                                wx.PENSTYLE_SOLID)
         self._tickLength = tuple(-x * 2 for x in self._pointSize)
 
         self._diagonalPen = wx.Pen(wx.BLUE,
-                                   self._pointSize[0],
+                                   int(self._pointSize[0]),
                                    wx.PENSTYLE_DOT_DASH)
 
     def SetCursor(self, cursor):
@@ -1920,10 +1920,10 @@ class PlotCanvas(wx.Panel):
         # set clipping area so drawing does not occur outside axis box
         ptx, pty, rectWidth, rectHeight = self._point2ClientCoord(p1, p2)
         # allow graph to overlap axis lines by adding units to w and h
-        dc.SetClippingRegion(ptx * self._pointSize[0],
-                             pty * self._pointSize[1],
-                             rectWidth * self._pointSize[0] + 2,
-                             rectHeight * self._pointSize[1] + 1)
+        dc.SetClippingRegion(int(ptx * self._pointSize[0]),
+                             int(pty * self._pointSize[1]),
+                             int(rectWidth * self._pointSize[0] + 2),
+                             int(rectHeight * self._pointSize[1] + 1))
         # Draw the lines and markers
 #        start = _time.perf_counter()
         graphics.draw(dc)
@@ -2275,7 +2275,7 @@ class PlotCanvas(wx.Panel):
             pnt = ((trhc[0] + legendLHS + legendSymExt[0]
                     + 5 * self._pointSize[0]),
                    trhc[1] + s + lineHeight / 2. - legendTextExt[1] / 2)
-            dc.DrawText(o.getLegend(), pnt[0], pnt[1])
+            dc.DrawText(o.getLegend(), int(pnt[0]), int(pnt[1]))
         dc.SetFont(self._getFont(self._fontSizeAxis))  # reset
 
     def _titleLablesWH(self, dc, graphics):
@@ -2324,7 +2324,7 @@ class PlotCanvas(wx.Panel):
         dc.SetPen(wx.Pen(wx.BLACK))
         dc.SetBrush(wx.Brush(wx.WHITE, wx.BRUSHSTYLE_TRANSPARENT))
         dc.SetLogicalFunction(wx.INVERT)
-        dc.DrawRectangle(ptx, pty, rectWidth, rectHeight)
+        dc.DrawRectangle(int(ptx), int(pty), int(rectWidth), int(rectHeight))
         dc.SetLogicalFunction(wx.COPY)
 
     def _getFont(self, size):
@@ -2418,7 +2418,7 @@ class PlotCanvas(wx.Panel):
         # increases thickness for printing only
         pen = self.gridPen
         penWidth = self.printerScale * pen.GetWidth()
-        pen.SetWidth(penWidth)
+        pen.SetWidth(int(penWidth))
         dc.SetPen(pen)
 
         x, y, width, height = self._point2ClientCoord(p1, p2)
@@ -2427,13 +2427,13 @@ class PlotCanvas(wx.Panel):
             if self.enableGrid[0]:
                 for x, _ in xticks:
                     pt = scale_and_shift_point(x, p1[1], scale, shift)
-                    dc.DrawLine(pt[0], pt[1], pt[0], pt[1] - height)
+                    dc.DrawLine(int(pt[0]), int(pt[1]), int(pt[0]), int(pt[1] - height))
 
         if self._ySpec != 'none':
             if self.enableGrid[1]:
                 for y, label in yticks:
                     pt = scale_and_shift_point(p1[0], y, scale, shift)
-                    dc.DrawLine(pt[0], pt[1], pt[0] + width, pt[1])
+                    dc.DrawLine(int(pt[0]), int(pt[1]), int(pt[0] + width), int(pt[1]))
 
     @TempStyle('pen')
     def _drawTicks(self, dc, p1, p2, scale, shift, xticks, yticks):
@@ -2466,7 +2466,7 @@ class PlotCanvas(wx.Panel):
         # increases thickness for printing only
         pen = self.tickPen
         penWidth = self.printerScale * pen.GetWidth()
-        pen.SetWidth(penWidth)
+        pen.SetWidth(int(penWidth))
         dc.SetPen(pen)
 
         # lengthen lines for printing
@@ -2479,13 +2479,13 @@ class PlotCanvas(wx.Panel):
                 lines = []
                 for x, label in xticks:
                     pt = scale_and_shift_point(x, p1[1], scale, shift)
-                    lines.append((pt[0], pt[1], pt[0], pt[1] - xTickLength))
+                    lines.append((int(pt[0]), int(pt[1]), int(pt[0]), int(pt[1] - xTickLength)))
                 dc.DrawLineList(lines)
             if ticks.top:
                 lines = []
                 for x, label in xticks:
                     pt = scale_and_shift_point(x, p2[1], scale, shift)
-                    lines.append((pt[0], pt[1], pt[0], pt[1] + xTickLength))
+                    lines.append((int(pt[0]), int(pt[1]), int(pt[0]), int(pt[1] + xTickLength)))
                 dc.DrawLineList(lines)
 
         if self.ySpec != 'none':
@@ -2493,13 +2493,13 @@ class PlotCanvas(wx.Panel):
                 lines = []
                 for y, label in yticks:
                     pt = scale_and_shift_point(p1[0], y, scale, shift)
-                    lines.append((pt[0], pt[1], pt[0] + yTickLength, pt[1]))
+                    lines.append((int(pt[0]), int(pt[1]), int(pt[0] + yTickLength), int(pt[1])))
                 dc.DrawLineList(lines)
             if ticks.right:
                 lines = []
                 for y, label in yticks:
                     pt = scale_and_shift_point(p2[0], y, scale, shift)
-                    lines.append((pt[0], pt[1], pt[0] - yTickLength, pt[1]))
+                    lines.append((int(pt[0]), int(pt[1]), int(pt[0] - yTickLength), int(pt[1])))
                 dc.DrawLineList(lines)
 
     @TempStyle('pen')
@@ -2526,25 +2526,25 @@ class PlotCanvas(wx.Panel):
         # increases thickness for printing only
         pen = self.centerLinePen
         penWidth = self.printerScale * pen.GetWidth()
-        pen.SetWidth(penWidth)
+        pen.SetWidth(int(penWidth))
         dc.SetPen(pen)
 
         if self._centerLinesEnabled in ('Horizontal', True):
             y1 = scale[1] * p1[1] + shift[1]
             y2 = scale[1] * p2[1] + shift[1]
             y = (y1 - y2) / 2.0 + y2
-            dc.DrawLine(scale[0] * p1[0] + shift[0],
-                        y,
-                        scale[0] * p2[0] + shift[0],
-                        y)
+            dc.DrawLine(int(scale[0] * p1[0] + shift[0]),
+                        int(y),
+                        int(scale[0] * p2[0] + shift[0]),
+                        int(y))
         if self._centerLinesEnabled in ('Vertical', True):
             x1 = scale[0] * p1[0] + shift[0]
             x2 = scale[0] * p2[0] + shift[0]
             x = (x1 - x2) / 2.0 + x2
-            dc.DrawLine(x,
-                        scale[1] * p1[1] + shift[1],
-                        x,
-                        scale[1] * p2[1] + shift[1])
+            dc.DrawLine(int(x),
+                        int(scale[1] * p1[1] + shift[1]),
+                        int(x),
+                        int(scale[1] * p2[1] + shift[1]))
 
     @TempStyle('pen')
     def _drawDiagonals(self, dc, p1, p2, scale, shift):
@@ -2570,19 +2570,19 @@ class PlotCanvas(wx.Panel):
         """
         pen = self.diagonalPen
         penWidth = self.printerScale * pen.GetWidth()
-        pen.SetWidth(penWidth)
+        pen.SetWidth(int(penWidth))
         dc.SetPen(pen)
 
         if self._diagonalsEnabled in ('Bottomleft-Topright', True):
-            dc.DrawLine(scale[0] * p1[0] + shift[0],
-                        scale[1] * p1[1] + shift[1],
-                        scale[0] * p2[0] + shift[0],
-                        scale[1] * p2[1] + shift[1])
+            dc.DrawLine(int(scale[0] * p1[0] + shift[0]),
+                        int(scale[1] * p1[1] + shift[1]),
+                        int(scale[0] * p2[0] + shift[0]),
+                        int(scale[1] * p2[1] + shift[1]))
         if self._diagonalsEnabled in ('Bottomright-Topleft', True):
-            dc.DrawLine(scale[0] * p1[0] + shift[0],
-                        scale[1] * p2[1] + shift[1],
-                        scale[0] * p2[0] + shift[0],
-                        scale[1] * p1[1] + shift[1])
+            dc.DrawLine(int(scale[0] * p1[0] + shift[0]),
+                        int(scale[1] * p2[1] + shift[1]),
+                        int(scale[0] * p2[0] + shift[0]),
+                        int(scale[1] * p1[1] + shift[1]))
 
     @TempStyle('pen')
     def _drawAxes(self, dc, p1, p2, scale, shift):
@@ -2609,7 +2609,7 @@ class PlotCanvas(wx.Panel):
         # increases thickness for printing only
         pen = self.axesPen
         penWidth = self.printerScale * pen.GetWidth()
-        pen.SetWidth(penWidth)
+        pen.SetWidth(int(penWidth))
         dc.SetPen(pen)
 
         axes = self.enableAxes
@@ -2618,24 +2618,24 @@ class PlotCanvas(wx.Panel):
                 lower, upper = p1[0], p2[0]
                 a1 = scale_and_shift_point(lower, p1[1], scale, shift)
                 a2 = scale_and_shift_point(upper, p1[1], scale, shift)
-                dc.DrawLine(a1[0], a1[1], a2[0], a2[1])
+                dc.DrawLine(int(a1[0]), int(a1[1]), int(a2[0]), int(a2[1]))
             if axes.top:
                 lower, upper = p1[0], p2[0]
                 a1 = scale_and_shift_point(lower, p2[1], scale, shift)
                 a2 = scale_and_shift_point(upper, p2[1], scale, shift)
-                dc.DrawLine(a1[0], a1[1], a2[0], a2[1])
+                dc.DrawLine(int(a1[0]), int(a1[1]), int(a2[0]), int(a2[1]))
 
         if self.ySpec != 'none':
             if axes.left:
                 lower, upper = p1[1], p2[1]
                 a1 = scale_and_shift_point(p1[0], lower, scale, shift)
                 a2 = scale_and_shift_point(p1[0], upper, scale, shift)
-                dc.DrawLine(a1[0], a1[1], a2[0], a2[1])
+                dc.DrawLine(int(a1[0]), int(a1[1]), int(a2[0]), int(a2[1]))
             if axes.right:
                 lower, upper = p1[1], p2[1]
                 a1 = scale_and_shift_point(p2[0], lower, scale, shift)
                 a2 = scale_and_shift_point(p2[0], upper, scale, shift)
-                dc.DrawLine(a1[0], a1[1], a2[0], a2[1])
+                dc.DrawLine(int(a1[0]), int(a1[1]), int(a2[0]), int(a2[1]))
 
     @TempStyle('pen')
     def _drawAxesValues(self, dc, p1, p2, scale, shift, xticks, yticks):
@@ -2681,8 +2681,8 @@ class PlotCanvas(wx.Panel):
                     w = dc.GetTextExtent(label)[0]
                     pt = scale_and_shift_point(x, p1[1], scale, shift)
                     coords.append(
-                        (pt[0] - w/2,
-                         pt[1] + 2 * self._pointSize[1] - xTickLength)
+                        (int(pt[0] - w/2),
+                         int(pt[1] + 2 * self._pointSize[1] - xTickLength))
                     )
                 dc.DrawTextList(labels, coords)
 
@@ -2693,8 +2693,8 @@ class PlotCanvas(wx.Panel):
                     w, h = dc.GetTextExtent(label)
                     pt = scale_and_shift_point(x, p2[1], scale, shift)
                     coords.append(
-                        (pt[0] - w/2,
-                         pt[1] - 2 * self._pointSize[1] - h - xTickLength)
+                        (int(pt[0] - w/2),
+                         int(pt[1] - 2 * self._pointSize[1] - h - xTickLength))
                     )
                 dc.DrawTextList(labels, coords)
 
@@ -2707,8 +2707,8 @@ class PlotCanvas(wx.Panel):
                     w = dc.GetTextExtent(label)[0]
                     pt = scale_and_shift_point(p1[0], y, scale, shift)
                     coords.append(
-                        (pt[0] - w - 3 * self._pointSize[0] + yTickLength,
-                         pt[1] - 0.5 * h)
+                        (int(pt[0] - w - 3 * self._pointSize[0] + yTickLength),
+                         int(pt[1] - 0.5 * h))
                     )
                 dc.DrawTextList(labels, coords)
 
@@ -2720,8 +2720,8 @@ class PlotCanvas(wx.Panel):
                     w = dc.GetTextExtent(label)[0]
                     pt = scale_and_shift_point(p2[0], y, scale, shift)
                     coords.append(
-                        (pt[0] + 3 * self._pointSize[0] + yTickLength,
-                         pt[1] - 0.5 * h)
+                        (int(pt[0] + 3 * self._pointSize[0] + yTickLength),
+                         int(pt[1] - 0.5 * h))
                     )
                 dc.DrawTextList(labels, coords)
 
@@ -2780,7 +2780,7 @@ class PlotCanvas(wx.Panel):
             + (self.plotbox_size[0] - lhsW - rhsW) / 2. - titleWH[0] / 2.,
             self.plotbox_origin[1] - self.plotbox_size[1]
         )
-        dc.DrawText(graphics.title, titlePos[0], titlePos[1])
+        dc.DrawText(graphics.title, int(titlePos[0]), int(titlePos[1]))
 
     def _drawAxesLabels(self, dc, graphics, lhsW, rhsW, bottomH, topH,
                         xLabelWH, yLabelWH):
@@ -2801,7 +2801,7 @@ class PlotCanvas(wx.Panel):
             + (self.plotbox_size[0] - lhsW - rhsW) / 2. - xLabelWH[0] / 2.,
             self.plotbox_origin[1] - xLabelWH[1] - yTickLength
         )
-        dc.DrawText(graphics.xLabel, xLabelPos[0], xLabelPos[1])
+        dc.DrawText(graphics.xLabel, int(xLabelPos[0]), int(xLabelPos[1]))
         yLabelPos = (
             self.plotbox_origin[0] - 3 * self._pointSize[0] + xTickLength,
             self.plotbox_origin[1] - bottomH
@@ -2809,7 +2809,7 @@ class PlotCanvas(wx.Panel):
         )
         if graphics.yLabel:  # bug fix for Linux
             dc.DrawRotatedText(
-                graphics.yLabel, yLabelPos[0], yLabelPos[1], 90)
+                graphics.yLabel, int(yLabelPos[0]), int(yLabelPos[1]), 90)
 
     @TempStyle('pen')
     def _drawPlotAreaLabels(self, dc, graphics, lhsW, rhsW, titleWH,
@@ -2936,7 +2936,7 @@ class PlotCanvas(wx.Panel):
         if pos >= 0:
             pagesize = int((r_current[1] - r_current[0]) / unit)
 
-            self.sb_hor.SetScrollbar(pos, pagesize, sbfullrange, pagesize)
+            self.sb_hor.SetScrollbar(pos, pagesize, int(sbfullrange), pagesize)
             self._sb_xunit = unit
             needScrollbars = needScrollbars or (pagesize != sbfullrange)
         else:
@@ -2958,7 +2958,7 @@ class PlotCanvas(wx.Panel):
         if pos >= 0:
             pagesize = int((r_current[1] - r_current[0]) / unit)
             pos = (sbfullrange - 1 - pos - pagesize)
-            self.sb_vert.SetScrollbar(pos, pagesize, sbfullrange, pagesize)
+            self.sb_vert.SetScrollbar(int(pos), pagesize, int(sbfullrange), pagesize)
             self._sb_yunit = unit
             needScrollbars = needScrollbars or (pagesize != sbfullrange)
         else:
